@@ -1,44 +1,43 @@
 module github.com/gruntwork-io/terragrunt
 
-go 1.20
+go 1.21
 
 require (
-	cloud.google.com/go/storage v1.27.0
-	github.com/aws/aws-sdk-go v1.44.122
+	cloud.google.com/go/storage v1.28.1
+	github.com/aws/aws-sdk-go v1.46.6
 	github.com/creack/pty v1.1.11
 	github.com/fatih/structs v1.1.0
-	github.com/go-errors/errors v1.0.2-0.20180813162953-d98b870cc4e0
+	github.com/go-errors/errors v1.4.2
 	github.com/google/shlex v0.0.0-20191202100458-e7afc7fbc510
-	github.com/gruntwork-io/terratest v0.32.6
+	github.com/gruntwork-io/terratest v0.41.0
 	github.com/hashicorp/go-cleanhttp v0.5.2
-	github.com/hashicorp/go-getter v1.7.0
+	github.com/hashicorp/go-getter v1.7.1
 	github.com/hashicorp/go-multierror v1.1.1
 	github.com/hashicorp/go-safetemp v1.0.0
 	github.com/hashicorp/go-version v1.6.0
-	github.com/hashicorp/hcl/v2 v2.15.0
+	github.com/hashicorp/hcl/v2 v2.17.0
 
 	// Many functions of terraform was converted to internal to avoid use as a library after v0.15.3. This means that we
 	// can't use terraform as a library after v0.15.3, so we pull that in here.
 	github.com/hashicorp/terraform v0.15.3
 	github.com/hashicorp/terraform-config-inspect v0.0.0-20210318070130-9a80970d6b34
 	github.com/imdario/mergo v0.3.11
-	github.com/mattn/go-zglob v0.0.2-0.20190814121620-e3c945676326
+	github.com/mattn/go-zglob v0.0.3
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/mitchellh/mapstructure v1.4.3
-	github.com/sirupsen/logrus v1.8.1
-	github.com/stretchr/testify v1.7.2
-	github.com/urfave/cli v1.22.3
-	github.com/zclconf/go-cty v1.12.1
+	github.com/sirupsen/logrus v1.9.3
+	github.com/stretchr/testify v1.8.4
+	github.com/zclconf/go-cty v1.13.2
 	go.mozilla.org/sops/v3 v3.7.3
-	golang.org/x/crypto v0.5.0
-	golang.org/x/oauth2 v0.1.0
-	golang.org/x/sync v0.0.0-20220929204114-8fcdb60fdcc0
-	golang.org/x/sys v0.5.0
-	google.golang.org/api v0.100.0
+	golang.org/x/crypto v0.14.0 // indirect
+	golang.org/x/oauth2 v0.8.0
+	golang.org/x/sync v0.3.0
+	golang.org/x/sys v0.13.0
+	google.golang.org/api v0.114.0
 )
 
 require (
-	cloud.google.com/go v0.104.0 // indirect
+	cloud.google.com/go v0.110.0 // indirect
 	github.com/Azure/go-autorest/autorest/azure/auth v0.5.11 // indirect
 	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
 	github.com/golang/snappy v0.0.4 // indirect
