@@ -1,7 +1,6 @@
 package configstack
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -29,7 +28,7 @@ func TestFindStackInSubfolders(t *testing.T) {
 	writeDummyTerragruntConfigs(t, tempFolder, filePaths)
 
 	envFolder := filepath.ToSlash(util.JoinPath(tempFolder + "/stage"))
-	terragruntOptions, err := options.NewTerragruntOptions(envFolder)
+	terragruntOptions, err := options.NewTerragruntOptionsWithConfigPath(envFolder)
 	if err != nil {
 		t.Fatalf("Failed when calling method under test: %s\n", err.Error())
 	}
@@ -153,7 +152,7 @@ func createTestStack() *Stack {
 }
 
 func createTempFolder(t *testing.T) string {
-	tmpFolder, err := ioutil.TempDir("", "")
+	tmpFolder, err := os.MkdirTemp("", "")
 	if err != nil {
 		t.Fatalf("Failed to create temp directory: %s\n", err.Error())
 	}
@@ -170,7 +169,7 @@ func writeDummyTerragruntConfigs(t *testing.T, tmpFolder string, paths []string)
 		containingDir := filepath.Dir(absPath)
 		createDirIfNotExist(t, containingDir)
 
-		err := ioutil.WriteFile(absPath, contents, os.ModePerm)
+		err := os.WriteFile(absPath, contents, os.ModePerm)
 		if err != nil {
 			t.Fatalf("Failed to write file at path %s: %s\n", path, err.Error())
 		}
