@@ -4,6 +4,9 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/aws/aws-sdk-go/aws/request"
+	"github.com/gruntwork-io/go-commons/version"
+
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/arn"
 	"github.com/aws/aws-sdk-go/aws/credentials"
@@ -11,7 +14,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/endpoints"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sts"
-	"github.com/gruntwork-io/terragrunt/errors"
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 )
 
@@ -27,6 +30,13 @@ type AwsSessionConfig struct {
 	DisableComputeChecksums bool
 	ExternalID              string
 	SessionName             string
+}
+
+// addUserAgent - Add terragrunt version to the user agent for AWS API calls.
+var addUserAgent = request.NamedHandler{
+	Name: "terragrunt.UserAgentHandler",
+	Fn: request.MakeAddToUserAgentHandler(
+		"terragrunt", version.GetVersion()),
 }
 
 // Returns an AWS session object for the given config region (required), profile name (optional), and IAM role to assume
@@ -70,6 +80,8 @@ func CreateAwsSessionFromConfig(config *AwsSessionConfig, terragruntOptions *opt
 	if err != nil {
 		return nil, errors.WithStackTraceAndPrefix(err, "Error initializing session")
 	}
+
+	sess.Handlers.Build.PushFrontNamed(addUserAgent)
 
 	// Merge the config based IAMRole options into the original one, as the config has higher precedence than CLI.
 	iamRoleOptions := terragruntOptions.IAMRoleOptions
@@ -125,6 +137,7 @@ func CreateAwsSession(config *AwsSessionConfig, terragruntOptions *options.Terra
 		if err != nil {
 			return nil, errors.WithStackTrace(err)
 		}
+		sess.Handlers.Build.PushFrontNamed(addUserAgent)
 		if terragruntOptions.IAMRoleOptions.RoleARN != "" {
 			terragruntOptions.Logger.Debugf("Assuming role %s", terragruntOptions.IAMRoleOptions.RoleARN)
 			sess.Config.Credentials = getSTSCredentialsFromIAMRoleOptions(sess, terragruntOptions.IAMRoleOptions)
@@ -155,6 +168,8 @@ func AssumeIamRole(iamRoleOpts options.IAMRoleOptions) (*sts.Credentials, error)
 	if err != nil {
 		return nil, errors.WithStackTrace(err)
 	}
+
+	sess.Handlers.Build.PushFrontNamed(addUserAgent)
 
 	_, err = sess.Config.Credentials.Get()
 	if err != nil {
@@ -200,6 +215,13 @@ func GetAWSCallerIdentity(config *AwsSessionConfig, terragruntOptions *options.T
 	}
 
 	return *identity, nil
+}
+
+// ValidateAwsSession - Validate if current AWS session is valid
+func ValidateAwsSession(config *AwsSessionConfig, terragruntOptions *options.TerragruntOptions) error {
+	// read the caller identity to check if the credentials are valid
+	_, err := GetAWSCallerIdentity(config, terragruntOptions)
+	return err
 }
 
 // Get the AWS Partition of the current session configuration
