@@ -13,12 +13,13 @@ type Policy struct {
 // https://docs.aws.amazon.com/IAM//latest/UserGuide/reference_policies_elements_action.html
 // https://docs.aws.amazon.com/IAM//latest/UserGuide/reference_policies_elements_resource.html
 type Statement struct {
-	Sid       string                  `json:"Sid"`
-	Effect    string                  `json:"Effect"`
-	Principal interface{}             `json:"Principal"`
-	Action    interface{}             `json:"Action"`
-	Resource  interface{}             `json:"Resource"`
-	Condition *map[string]interface{} `json:"Condition,omitempty"`
+	Sid          string                  `json:"Sid"`
+	Effect       string                  `json:"Effect"`
+	Principal    interface{}             `json:"Principal,omitempty"`
+	NotPrincipal interface{}             `json:"NotPrincipal,omitempty"`
+	Action       interface{}             `json:"Action"`
+	Resource     interface{}             `json:"Resource"`
+	Condition    *map[string]interface{} `json:"Condition,omitempty"`
 }
 
 func UnmarshalPolicy(policy string) (Policy, error) {
