@@ -1,7 +1,9 @@
 package config
 
 import (
+	"encoding/json"
 	"fmt"
+	"os"
 	"path/filepath"
 	"regexp"
 	"runtime"
@@ -17,12 +19,15 @@ import (
 	"github.com/zclconf/go-cty/cty/function"
 	"go.mozilla.org/sops/v3/decrypt"
 
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/aws_helper"
-	"github.com/gruntwork-io/terragrunt/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/shell"
 	"github.com/gruntwork-io/terragrunt/util"
 )
+
+const noMatchedPats = 1
+const matchedPats = 2
 
 // List of terraform commands that accept -lock-timeout
 var TERRAFORM_COMMANDS_NEED_LOCKING = []string{
@@ -140,6 +145,8 @@ func CreateTerragruntEvalContext(
 		"get_terraform_commands_that_need_parallelism": wrapStaticValueToStringSliceAsFuncImpl(TERRAFORM_COMMANDS_NEED_PARALLELISM),
 		"sops_decrypt_file":                            wrapStringSliceToStringAsFuncImpl(sopsDecryptFile, extensions.TrackInclude, terragruntOptions),
 		"get_terragrunt_source_cli_flag":               wrapVoidToStringAsFuncImpl(getTerragruntSourceCliFlag, extensions.TrackInclude, terragruntOptions),
+		"get_default_retryable_errors":                 wrapVoidToStringSliceAsFuncImpl(getDefaultRetryableErrors, extensions.TrackInclude, terragruntOptions),
+		"read_tfvars_file":                             wrapStringSliceToStringAsFuncImpl(readTFVarsFile, extensions.TrackInclude, terragruntOptions),
 	}
 
 	// Map with HCL functions introduced in Terraform after v0.15.3, since upgrade to a later version is not supported
@@ -216,12 +223,12 @@ func getPathToRepoRoot(trackInclude *TrackInclude, terragruntOptions *options.Te
 		return "", errors.WithStackTrace(err)
 	}
 
-	repoRootPathAbs, err := filepath.Rel(terragruntOptions.WorkingDir, string(repoAbsPath))
+	repoRootPathAbs, err := filepath.Rel(terragruntOptions.WorkingDir, repoAbsPath)
 	if err != nil {
 		return "", errors.WithStackTrace(err)
 	}
 
-	return filepath.ToSlash(strings.TrimSpace(repoRootPathAbs) + "/"), nil
+	return filepath.ToSlash(strings.TrimSpace(repoRootPathAbs)), nil
 }
 
 // Return the directory where the Terragrunt configuration file lives
@@ -267,10 +274,10 @@ func parseGetEnvParameters(parameters []string) (EnvVar, error) {
 	envVariable := EnvVar{}
 
 	switch len(parameters) {
-	case 1:
+	case noMatchedPats:
 		envVariable.IsRequired = true
 		envVariable.Name = parameters[0]
-	case 2:
+	case matchedPats:
 		envVariable.Name = parameters[0]
 		envVariable.DefaultValue = parameters[1]
 	default:
@@ -381,7 +388,7 @@ func findInParentFolders(
 	if numParams > 1 {
 		fallbackParam = params[1]
 	}
-	if numParams > 2 {
+	if numParams > matchedPats {
 		return "", errors.WithStackTrace(WrongNumberOfParams{Func: "find_in_parent_folders", Expected: "0, 1, or 2", Actual: numParams})
 	}
 
@@ -402,7 +409,7 @@ func findInParentFolders(
 	for i := 0; i < terragruntOptions.MaxFoldersToCheck; i++ {
 		currentDir := filepath.ToSlash(filepath.Dir(previousDir))
 		if currentDir == previousDir {
-			if numParams == 2 {
+			if numParams == matchedPats {
 				return fallbackParam, nil
 			}
 			return "", errors.WithStackTrace(ParentFileNotFound{Path: terragruntOptions.TerragruntConfigPath, File: fileToFindStr, Cause: "Traversed all the way to the root"})
@@ -432,16 +439,17 @@ func pathRelativeToInclude(params []string, trackInclude *TrackInclude, terragru
 	}
 
 	var included IncludeConfig
-	if trackInclude.Original != nil {
+	switch {
+	case trackInclude.Original != nil:
 		included = *trackInclude.Original
-	} else if len(trackInclude.CurrentList) > 0 {
+	case len(trackInclude.CurrentList) > 0:
 		// Called in child context, so we need to select the right include file.
 		selected, err := getSelectedIncludeBlock(*trackInclude, params)
 		if err != nil {
 			return "", err
 		}
 		included = *selected
-	} else {
+	default:
 		return ".", nil
 	}
 
@@ -487,6 +495,11 @@ func getTerraformCommand(trackInclude *TrackInclude, terragruntOptions *options.
 // getTerraformCliArgs returns cli args for terraform
 func getTerraformCliArgs(trackInclude *TrackInclude, terragruntOptions *options.TerragruntOptions) ([]string, error) {
 	return terragruntOptions.TerraformCliArgs, nil
+}
+
+// getDefaultRetryableErrors returns default retryable errors
+func getDefaultRetryableErrors(trackInclude *TrackInclude, terragruntOptions *options.TerragruntOptions) ([]string, error) {
+	return options.DEFAULT_RETRYABLE_ERRORS, nil
 }
 
 // Return the AWS account id associated to the current set of credentials
@@ -542,7 +555,10 @@ func readTerragruntConfig(configPath string, defaultVal *cty.Value, terragruntOp
 	// NOTE: this will not call terragrunt output, since all the values are cached from the ParseConfigFile call
 	// NOTE: we don't use range here because range will copy the slice, thereby undoing the set attribute.
 	for i := 0; i < len(config.TerragruntDependencies); i++ {
-		config.TerragruntDependencies[i].setRenderedOutputs(targetOptions)
+		err := config.TerragruntDependencies[i].setRenderedOutputs(targetOptions)
+		if err != nil {
+			return cty.NilVal, errors.WithStackTrace(err)
+		}
 	}
 
 	return TerragruntConfigAsCty(config)
@@ -569,7 +585,7 @@ func readTerragruntConfigAsFuncImpl(terragruntOptions *options.TerragruntOptions
 			}
 
 			var defaultVal *cty.Value = nil
-			if numParams == 2 {
+			if numParams == matchedPats {
 				defaultVal = &args[1]
 			}
 
@@ -651,7 +667,7 @@ func getModulePathFromSourceUrl(sourceUrl string) (string, error) {
 	matches := moduleNameRegexp.FindStringSubmatch(sourceUrl)
 
 	// if regexp returns less/more than the full match + 1 capture group, then something went wrong with regex (invalid source string)
-	if len(matches) != 2 {
+	if len(matches) != matchedPats {
 		return "", errors.WithStackTrace(ErrorParsingModulePath{ModuleSourceUrl: sourceUrl})
 	}
 
@@ -797,7 +813,7 @@ func endsWith(args []string, trackInclude *TrackInclude, terragruntOptions *opti
 
 // timeCmp implements Terraform's `timecmp` function that compares two timestamps.
 func timeCmp(args []string, trackInclude *TrackInclude, terragruntOptions *options.TerragruntOptions) (int64, error) {
-	if len(args) != 2 {
+	if len(args) != matchedPats {
 		return 0, errors.WithStackTrace(fmt.Errorf("function can take only two parameters: timestamp_a and timestamp_b"))
 	}
 
@@ -836,7 +852,61 @@ func strContains(args []string, trackInclude *TrackInclude, terragruntOptions *o
 	return false, nil
 }
 
+// readTFVarsFile reads a *.tfvars or *.tfvars.json file and returns the contents as a JSON encoded string
+func readTFVarsFile(args []string, trackInclude *TrackInclude, terragruntOptions *options.TerragruntOptions) (string, error) {
+
+	if len(args) != 1 {
+		return "", errors.WithStackTrace(WrongNumberOfParams{Func: "read_tfvars_file", Expected: "1", Actual: len(args)})
+	}
+
+	varFile := args[0]
+	varFile, err := util.CanonicalPath(varFile, terragruntOptions.WorkingDir)
+	if err != nil {
+		return "", errors.WithStackTrace(err)
+	}
+
+	if !util.FileExists(varFile) {
+		return "", errors.WithStackTrace(TFVarFileNotFoundError{File: varFile})
+	}
+
+	fileContents, err := os.ReadFile(varFile)
+	if err != nil {
+		return "", errors.WithStackTrace(fmt.Errorf("could not read file %q: %w", varFile, err))
+	}
+
+	if strings.HasSuffix(varFile, "json") {
+		var variables map[string]interface{}
+		// just want to be sure that the file is valid json
+		if err := json.Unmarshal(fileContents, &variables); err != nil {
+			return "", errors.WithStackTrace(fmt.Errorf("could not unmarshal json body of tfvar file: %w", err))
+		}
+		return string(fileContents), nil
+	}
+
+	var variables map[string]interface{}
+	if err := ParseAndDecodeVarFile(string(fileContents), varFile, &variables); err != nil {
+		return "", err
+	}
+
+	data, err := json.Marshal(variables)
+	if err != nil {
+		return "", errors.WithStackTrace(fmt.Errorf("could not marshal json body of tfvar file: %w", err))
+	}
+
+	return string(data), nil
+}
+
 // Custom error types
+
+type TFVarFileNotFoundError struct {
+	File  string
+	Cause string
+}
+
+func (err TFVarFileNotFoundError) Error() string {
+	return fmt.Sprintf("TFVarFileNotFound: Could not find a %s. Cause: %s.", err.File, err.Cause)
+}
+
 type WrongNumberOfParams struct {
 	Func     string
 	Expected string
