@@ -3,18 +3,20 @@ package configstack
 import (
 	"encoding/json"
 	"fmt"
+	"os"
 	"path/filepath"
 	"sort"
 	"strings"
 
 	"github.com/sirupsen/logrus"
 
+	"github.com/gruntwork-io/go-commons/collections"
+	"github.com/gruntwork-io/go-commons/errors"
+	"github.com/gruntwork-io/go-commons/files"
 	"github.com/gruntwork-io/terragrunt/config"
-	"github.com/gruntwork-io/terragrunt/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/shell"
 	"github.com/gruntwork-io/terragrunt/util"
-	zglob "github.com/mattn/go-zglob"
 )
 
 const maxLevelsOfRecursion = 20
@@ -51,154 +53,70 @@ func (module TerraformModule) MarshalJSON() ([]byte, error) {
 func ResolveTerraformModules(terragruntConfigPaths []string, terragruntOptions *options.TerragruntOptions, childTerragruntConfig *config.TerragruntConfig, howThesePathsWereFound string) ([]*TerraformModule, error) {
 	canonicalTerragruntConfigPaths, err := util.CanonicalPaths(terragruntConfigPaths, ".")
 	if err != nil {
-		return []*TerraformModule{}, err
+		return nil, err
 	}
 
 	modules, err := resolveModules(canonicalTerragruntConfigPaths, terragruntOptions, childTerragruntConfig, howThesePathsWereFound)
 	if err != nil {
-		return []*TerraformModule{}, err
+		return nil, err
 	}
 
 	externalDependencies, err := resolveExternalDependenciesForModules(modules, map[string]*TerraformModule{}, 0, terragruntOptions, childTerragruntConfig)
 	if err != nil {
-		return []*TerraformModule{}, err
+		return nil, err
 	}
 
 	crossLinkedModules, err := crosslinkDependencies(mergeMaps(modules, externalDependencies), canonicalTerragruntConfigPaths)
 	if err != nil {
-		return []*TerraformModule{}, err
+		return nil, err
 	}
 
-	includedModules, err := flagIncludedDirs(crossLinkedModules, terragruntOptions)
-	if err != nil {
-		return []*TerraformModule{}, err
-	}
+	includedModules := flagIncludedDirs(crossLinkedModules, terragruntOptions)
 
-	includedModulesWithExcluded, err := flagExcludedDirs(includedModules, terragruntOptions)
-	if err != nil {
-		return []*TerraformModule{}, err
-	}
+	includedModulesWithExcluded := flagExcludedDirs(includedModules, terragruntOptions)
 
 	finalModules, err := flagModulesThatDontInclude(includedModulesWithExcluded, terragruntOptions)
 	if err != nil {
-		return []*TerraformModule{}, err
+		return nil, err
 	}
 
 	return finalModules, nil
 }
 
 // flagExcludedDirs iterates over a module slice and flags all entries as excluded, which should be ignored via the terragrunt-exclude-dir CLI flag.
-func flagExcludedDirs(modules []*TerraformModule, terragruntOptions *options.TerragruntOptions) ([]*TerraformModule, error) {
-
-	// If no ExcludeDirs is specified return the modules list instantly
-	if len(terragruntOptions.ExcludeDirs) == 0 {
-		return modules, nil
-	}
-
-	canonicalWorkingDir, err := util.CanonicalPath("", terragruntOptions.WorkingDir)
-	if err != nil {
-		return nil, err
-	}
-
-	excludeGlobMatches := []string{}
-
-	// If possible, expand the glob to get all excluded filepaths
-	for _, dir := range terragruntOptions.ExcludeDirs {
-
-		absoluteDir := ""
-
-		// Ensure excludedDirs are absolute
-		if filepath.IsAbs(dir) {
-			absoluteDir = dir
-		} else {
-			absoluteDir = filepath.Join(canonicalWorkingDir, dir)
-		}
-
-		matches, err := zglob.Glob(absoluteDir)
-
-		// Skip globs that can not be expanded
-		if err == nil {
-			excludeGlobMatches = append(excludeGlobMatches, matches...)
-		}
-	}
-
-	// Make sure all paths are canonical
-	canonicalExcludeDirs := []string{}
-	for _, module := range excludeGlobMatches {
-		canonicalPath, err := util.CanonicalPath(module, terragruntOptions.WorkingDir)
-		if err != nil {
-			return nil, err
-		}
-		canonicalExcludeDirs = append(canonicalExcludeDirs, canonicalPath)
-	}
-
+func flagExcludedDirs(modules []*TerraformModule, terragruntOptions *options.TerragruntOptions) []*TerraformModule {
 	for _, module := range modules {
-		if findModuleinPath(module, canonicalExcludeDirs) {
+		if findModuleInPath(module, terragruntOptions.ExcludeDirs) {
 			// Mark module itself as excluded
 			module.FlagExcluded = true
 		}
 
 		// Mark all affected dependencies as excluded
 		for _, dependency := range module.Dependencies {
-			if findModuleinPath(dependency, canonicalExcludeDirs) {
+			if findModuleInPath(dependency, terragruntOptions.ExcludeDirs) {
 				dependency.FlagExcluded = true
 			}
 		}
 	}
 
-	return modules, nil
+	return modules
 }
 
-// flagIncludedDirs iterates over a module slice and flags all entries not in the list specified via the terragrunt-include-dir CLI flag  as excluded.
-func flagIncludedDirs(modules []*TerraformModule, terragruntOptions *options.TerragruntOptions) ([]*TerraformModule, error) {
+// flagIncludedDirs iterates over a module slice and flags all entries not in the list specified via the terragrunt-include-dir CLI flag as excluded.
+func flagIncludedDirs(modules []*TerraformModule, terragruntOptions *options.TerragruntOptions) []*TerraformModule {
 
 	// If no IncludeDirs is specified return the modules list instantly
 	if len(terragruntOptions.IncludeDirs) == 0 {
 		// If we aren't given any include directories, but are given the strict include flag,
 		// return no modules.
 		if terragruntOptions.StrictInclude {
-			return []*TerraformModule{}, nil
+			return []*TerraformModule{}
 		}
-		return modules, nil
-	}
-
-	canonicalWorkingDir, err := util.CanonicalPath("", terragruntOptions.WorkingDir)
-	if err != nil {
-		return nil, err
-	}
-
-	includeGlobMatches := []string{}
-
-	// If possible, expand the glob to get all included filepaths
-	for _, dir := range terragruntOptions.IncludeDirs {
-
-		absoluteDir := dir
-
-		// Ensure includedDirs are absolute
-		if !filepath.IsAbs(dir) {
-			absoluteDir = filepath.Join(canonicalWorkingDir, dir)
-		}
-
-		matches, err := zglob.Glob(absoluteDir)
-
-		// Skip globs that can not be expanded
-		if err == nil {
-			includeGlobMatches = append(includeGlobMatches, matches...)
-		}
-	}
-
-	// Make sure all paths are canonical
-	canonicalIncludeDirs := []string{}
-	for _, module := range includeGlobMatches {
-		canonicalPath, err := util.CanonicalPath(module, terragruntOptions.WorkingDir)
-		if err != nil {
-			return nil, err
-		}
-		canonicalIncludeDirs = append(canonicalIncludeDirs, canonicalPath)
+		return modules
 	}
 
 	for _, module := range modules {
-		if findModuleinPath(module, canonicalIncludeDirs) {
+		if findModuleInPath(module, terragruntOptions.IncludeDirs) {
 			module.FlagExcluded = false
 		} else {
 			module.FlagExcluded = true
@@ -216,11 +134,11 @@ func flagIncludedDirs(modules []*TerraformModule, terragruntOptions *options.Ter
 		}
 	}
 
-	return modules, nil
+	return modules
 }
 
-// Returns true if a module is located under one of the target directories
-func findModuleinPath(module *TerraformModule, targetDirs []string) bool {
+// findModuleInPath returns true if a module is located under one of the target directories
+func findModuleInPath(module *TerraformModule, targetDirs []string) bool {
 	for _, targetDir := range targetDirs {
 		if module.Path == targetDir {
 			return true
@@ -301,12 +219,19 @@ func resolveModules(canonicalTerragruntConfigPaths []string, terragruntOptions *
 	moduleMap := map[string]*TerraformModule{}
 
 	for _, terragruntConfigPath := range canonicalTerragruntConfigPaths {
-		module, err := resolveTerraformModule(terragruntConfigPath, terragruntOptions, childTerragruntConfig, howTheseModulesWereFound)
+		module, err := resolveTerraformModule(terragruntConfigPath, moduleMap, terragruntOptions, childTerragruntConfig, howTheseModulesWereFound)
 		if err != nil {
 			return moduleMap, err
 		}
+
 		if module != nil {
 			moduleMap[module.Path] = module
+
+			dependencies, err := resolveDependenciesForModule(module, moduleMap, terragruntOptions, childTerragruntConfig, true)
+			if err != nil {
+				return moduleMap, err
+			}
+			moduleMap = collections.MergeMaps(moduleMap, dependencies)
 		}
 	}
 
@@ -316,10 +241,14 @@ func resolveModules(canonicalTerragruntConfigPaths []string, terragruntOptions *
 // Create a TerraformModule struct for the Terraform module specified by the given Terragrunt configuration file path.
 // Note that this method will NOT fill in the Dependencies field of the TerraformModule struct (see the
 // crosslinkDependencies method for that).
-func resolveTerraformModule(terragruntConfigPath string, terragruntOptions *options.TerragruntOptions, childTerragruntConfig *config.TerragruntConfig, howThisModuleWasFound string) (*TerraformModule, error) {
+func resolveTerraformModule(terragruntConfigPath string, moduleMap map[string]*TerraformModule, terragruntOptions *options.TerragruntOptions, childTerragruntConfig *config.TerragruntConfig, howThisModuleWasFound string) (*TerraformModule, error) {
 	modulePath, err := util.CanonicalPath(filepath.Dir(terragruntConfigPath), ".")
 	if err != nil {
 		return nil, err
+	}
+
+	if _, ok := moduleMap[modulePath]; ok {
+		return nil, nil
 	}
 
 	// Clone the options struct so we don't modify the original one. This is especially important as run-all operations
@@ -337,6 +266,11 @@ func resolveTerraformModule(terragruntConfigPath string, terragruntOptions *opti
 	if childTerragruntConfig != nil && childTerragruntConfig.ProcessedIncludes.ContainsPath(terragruntConfigPath) {
 		includeConfig = &config.IncludeConfig{Path: terragruntConfigPath}
 		opts.TerragruntConfigPath = terragruntOptions.OriginalTerragruntConfigPath
+	}
+
+	if collections.ListContainsElement(opts.ExcludeDirs, modulePath) {
+		// module is excluded
+		return &TerraformModule{Path: modulePath, TerragruntOptions: opts, FlagExcluded: true}, nil
 	}
 
 	// We only partially parse the config, only using the pieces that we need in this section. This config will be fully
@@ -416,7 +350,7 @@ func resolveExternalDependenciesForModules(moduleMap map[string]*TerraformModule
 	sortedKeys := getSortedKeys(moduleMap)
 	for _, key := range sortedKeys {
 		module := moduleMap[key]
-		externalDependencies, err := resolveExternalDependenciesForModule(module, modulesToSkip, terragruntOptions, childTerragruntConfig)
+		externalDependencies, err := resolveDependenciesForModule(module, modulesToSkip, terragruntOptions, childTerragruntConfig, false)
 		if err != nil {
 			return externalDependencies, err
 		}
@@ -450,12 +384,10 @@ func resolveExternalDependenciesForModules(moduleMap map[string]*TerraformModule
 	return allExternalDependencies, nil
 }
 
-// Look through the dependencies of the given module and resolve the "external" dependency paths listed in the module's
-// config (i.e. those dependencies not in the given list of Terragrunt config canonical file paths). These external
-// dependencies are outside of the current working directory, which means they may not be part of the environment the
-// user is trying to apply-all or destroy-all. Note that this method will NOT fill in the Dependencies field of the
-// TerraformModule struct (see the crosslinkDependencies method for that).
-func resolveExternalDependenciesForModule(module *TerraformModule, moduleMap map[string]*TerraformModule, terragruntOptions *options.TerragruntOptions, chilTerragruntConfig *config.TerragruntConfig) (map[string]*TerraformModule, error) {
+// resolveDependenciesForModule looks through the dependencies of the given module and resolve the dependency paths listed in the module's config.
+// If `skipExternal` is true, the func returns only dependencies that are inside of the current working directory, which means they are part of the environment the
+// user is trying to apply-all or destroy-all. Note that this method will NOT fill in the Dependencies field of the TerraformModule struct (see the crosslinkDependencies method for that).
+func resolveDependenciesForModule(module *TerraformModule, moduleMap map[string]*TerraformModule, terragruntOptions *options.TerragruntOptions, chilTerragruntConfig *config.TerragruntConfig, skipExternal bool) (map[string]*TerraformModule, error) {
 	if module.Config.Dependencies == nil || len(module.Config.Dependencies.Paths) == 0 {
 		return map[string]*TerraformModule{}, nil
 	}
@@ -467,7 +399,12 @@ func resolveExternalDependenciesForModule(module *TerraformModule, moduleMap map
 			return map[string]*TerraformModule{}, err
 		}
 
+		if skipExternal && !util.HasPathPrefix(dependencyPath, terragruntOptions.WorkingDir) {
+			continue
+		}
+
 		terragruntConfigPath := config.GetDefaultConfigPath(dependencyPath)
+
 		if _, alreadyContainsModule := moduleMap[dependencyPath]; !alreadyContainsModule {
 			externalTerragruntConfigPaths = append(externalTerragruntConfigPaths, terragruntConfigPath)
 		}
@@ -553,6 +490,10 @@ func getDependenciesForModule(module *TerraformModule, moduleMap map[string]*Ter
 			return dependencies, nil
 		}
 
+		if files.FileExists(dependencyModulePath) && !files.IsDir(dependencyModulePath) {
+			dependencyModulePath = filepath.Dir(dependencyModulePath)
+		}
+
 		dependencyModule, foundModule := moduleMap[dependencyModulePath]
 		if !foundModule {
 			err := UnrecognizedDependency{
@@ -590,9 +531,12 @@ func FindWhereWorkingDirIsIncluded(terragruntOptions *options.TerragruntOptions,
 	var matchedModulesMap = make(map[string]*TerraformModule)
 
 	gitTopLevelDir, err := shell.GitTopLevelDir(terragruntOptions, terragruntOptions.WorkingDir)
-	if err == nil { // top level detection worked
-		pathsToCheck = append(pathsToCheck, gitTopLevelDir)
-	} else { // detection failed, trying to use include directories as source for stacks
+	useIncludes := err != nil // fallback to includes git top level directory detection failed
+	if err == nil {
+		pathsToCheck, err = buildDirList(terragruntOptions, gitTopLevelDir)
+		useIncludes = err != nil // fallback to includes if directory list building failed
+	}
+	if useIncludes { // detection failed, trying to use include directories as source for stacks
 		uniquePaths := make(map[string]bool)
 		for _, includePath := range terragruntConfig.ProcessedIncludes {
 			uniquePaths[filepath.Dir(includePath.Path)] = true
@@ -603,28 +547,28 @@ func FindWhereWorkingDirIsIncluded(terragruntOptions *options.TerragruntOptions,
 	}
 
 	for _, dir := range pathsToCheck { // iterate over detected paths, build stacks and filter modules by working dir
-		dir = dir + filepath.FromSlash("/")
-		cfgOptions, err := options.NewTerragruntOptions(dir)
+		dir += filepath.FromSlash("/")
+		cfgOptions, err := options.NewTerragruntOptionsWithConfigPath(dir)
 		if err != nil {
 			terragruntOptions.Logger.Debugf("Failed to build terragrunt options from %s %v", dir, err)
-			return nil
+			continue
 		}
 
 		cfgOptions.Env = terragruntOptions.Env
 		cfgOptions.LogLevel = terragruntOptions.LogLevel
 		cfgOptions.OriginalTerragruntConfigPath = terragruntOptions.OriginalTerragruntConfigPath
+		cfgOptions.TerraformCommand = terragruntOptions.TerraformCommand
+		cfgOptions.NonInteractive = true
 
-		if terragruntOptions.TerraformCommand == "destroy" {
-			var hook = NewForceLogLevelHook(logrus.DebugLevel)
-			cfgOptions.Logger.Logger.AddHook(hook)
-		}
+		var hook = NewForceLogLevelHook(logrus.DebugLevel)
+		cfgOptions.Logger.Logger.AddHook(hook)
 
 		stack, err := FindStackInSubfolders(cfgOptions, terragruntConfig)
 		if err != nil {
-			// loggign error as debug since in some cases stack building may fail because parent files can be designed
+			// log error as debug since in some cases stack building may fail because parent files can be designed
 			// to work with relative paths from downstream modules
 			terragruntOptions.Logger.Debugf("Failed to build module stack %v", err)
-			return nil
+			continue
 		}
 
 		for _, module := range stack.Modules {
@@ -644,37 +588,6 @@ func FindWhereWorkingDirIsIncluded(terragruntOptions *options.TerragruntOptions,
 	}
 
 	return matchedModules
-}
-
-// Custom error types
-
-type UnrecognizedDependency struct {
-	ModulePath            string
-	DependencyPath        string
-	TerragruntConfigPaths []string
-}
-
-func (err UnrecognizedDependency) Error() string {
-	return fmt.Sprintf("Module %s specifies %s as a dependency, but that dependency was not one of the ones found while scanning subfolders: %v", err.ModulePath, err.DependencyPath, err.TerragruntConfigPaths)
-}
-
-type ErrorProcessingModule struct {
-	UnderlyingError       error
-	ModulePath            string
-	HowThisModuleWasFound string
-}
-
-func (err ErrorProcessingModule) Error() string {
-	return fmt.Sprintf("Error processing module at '%s'. How this module was found: %s. Underlying error: %v", err.ModulePath, err.HowThisModuleWasFound, err.UnderlyingError)
-}
-
-type InfiniteRecursion struct {
-	RecursionLevel int
-	Modules        map[string]*TerraformModule
-}
-
-func (err InfiniteRecursion) Error() string {
-	return fmt.Sprintf("Hit what seems to be an infinite recursion after going %d levels deep. Please check for a circular dependency! Modules involved: %v", err.RecursionLevel, err.Modules)
 }
 
 // ForceLogLevelHook - log hook which can change log level for messages which contains specific substrings
@@ -716,4 +629,22 @@ func (formatter *LogEntriesDropperFormatter) Format(entry *logrus.Entry) ([]byte
 		return formatter.OriginalFormatter.Format(entry)
 	}
 	return []byte(""), nil
+}
+
+// buildDirList - build list of directories from working directory to git top level directory
+func buildDirList(terragruntOptions *options.TerragruntOptions, topLevelDir string) ([]string, error) {
+	var pathsToCheck []string
+	relativePath, err := util.GetPathRelativeTo(topLevelDir, terragruntOptions.WorkingDir)
+	if err != nil {
+		return pathsToCheck, err
+	}
+	// build list of directories from working directory to git top level directory
+	// from which later will be built stacks
+	pathToAdd := terragruntOptions.WorkingDir
+	splits := strings.Split(relativePath, string(os.PathSeparator))
+	for _, path := range splits {
+		pathToAdd = filepath.Join(pathToAdd, path)
+		pathsToCheck = append(pathsToCheck, pathToAdd)
+	}
+	return pathsToCheck, nil
 }
