@@ -7,15 +7,17 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/aws_helper"
-	"github.com/gruntwork-io/terragrunt/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/util"
 )
 
 // DynamoDB only allows 10 table creates/deletes simultaneously. To ensure we don't hit this error, especially when
 // running many automated tests in parallel, we use a counting semaphore
-var tableCreateDeleteSemaphore = NewCountingSemaphore(10)
+const dynamoParallelOperations = 10
+
+var tableCreateDeleteSemaphore = NewCountingSemaphore(dynamoParallelOperations)
 
 // Terraform requires the DynamoDB table to have a primary key with this name
 const ATTR_LOCK_ID = "LockID"
@@ -25,6 +27,9 @@ const MAX_RETRIES_WAITING_FOR_TABLE_TO_BE_ACTIVE = 30
 const SLEEP_BETWEEN_TABLE_STATUS_CHECKS = 10 * time.Second
 
 const DYNAMODB_PAY_PER_REQUEST_BILLING_MODE = "PAY_PER_REQUEST"
+
+const sleepBetweenRetries = 20 * time.Second
+const maxRetries = 15
 
 // Create an authenticated client for DynamoDB
 func CreateDynamoDbClient(config *aws_helper.AwsSessionConfig, terragruntOptions *options.TerragruntOptions) (*dynamodb.DynamoDB, error) {
@@ -127,7 +132,7 @@ func CreateLockTable(tableName string, tags map[string]string, client *dynamodb.
 
 func tagTableIfTagsGiven(tags map[string]string, tableArn *string, client *dynamodb.DynamoDB, terragruntOptions *options.TerragruntOptions) error {
 
-	if tags == nil || len(tags) == 0 {
+	if len(tags) == 0 {
 		terragruntOptions.Logger.Debugf("No tags for lock table given.")
 		return nil
 	}
@@ -237,9 +242,6 @@ func UpdateLockTableSetSSEncryptionOnIfNecessary(tableName string, client *dynam
 
 // Wait until encryption is enabled for the given table
 func waitForEncryptionToBeEnabled(tableName string, client *dynamodb.DynamoDB, terragruntOptions *options.TerragruntOptions) error {
-	maxRetries := 15
-	sleepBetweenRetries := 20 * time.Second
-
 	terragruntOptions.Logger.Debugf("Waiting for encryption to be enabled on table %s", tableName)
 
 	for i := 0; i < maxRetries; i++ {
