@@ -3,25 +3,17 @@ package main
 import (
 	"os"
 
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/cli"
-	"github.com/gruntwork-io/terragrunt/errors"
 	"github.com/gruntwork-io/terragrunt/shell"
 	"github.com/gruntwork-io/terragrunt/util"
 )
 
-// This variable is set at build time using -ldflags parameters. For more info, see:
-// http://stackoverflow.com/a/11355611/483528
-var VERSION string
-
 // The main entrypoint for Terragrunt
 func main() {
-	// Log the terragrunt version in debug mode. This helps with debugging issues and ensuring a specific version of
-	// terragrunt used.
-	util.GlobalFallbackLogEntry.Debugf("Terragrunt Version: %s", VERSION)
-
 	defer errors.Recover(checkForErrorsAndExit)
 
-	app := cli.CreateTerragruntCli(VERSION, os.Stdout, os.Stderr)
+	app := cli.NewApp(os.Stdout, os.Stderr)
 	err := app.Run(os.Args)
 
 	checkForErrorsAndExit(err)
