@@ -9,8 +9,10 @@ import (
 	"strconv"
 	"time"
 
+	"google.golang.org/api/impersonate"
+
 	"cloud.google.com/go/storage"
-	"github.com/gruntwork-io/terragrunt/errors"
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/shell"
 	"github.com/gruntwork-io/terragrunt/util"
@@ -71,6 +73,11 @@ type accountFile struct {
 
 const MAX_RETRIES_WAITING_FOR_GCS_BUCKET = 12
 const SLEEP_BETWEEN_RETRIES_WAITING_FOR_GCS_BUCKET = 5 * time.Second
+
+const (
+	gcpMaxRetries          = 3
+	gcpSleepBetweenRetries = 10 * time.Second
+)
 
 type GCSInitializer struct{}
 
@@ -161,7 +168,7 @@ func (gcsInitializer GCSInitializer) Initialize(remoteState *RemoteState, terrag
 		return err
 	}
 
-	if err := validateGCSConfig(gcsConfigExtended, terragruntOptions); err != nil {
+	if err := validateGCSConfig(gcsConfigExtended); err != nil {
 		return err
 	}
 
@@ -232,11 +239,11 @@ func parseExtendedGCSConfig(config map[string]interface{}) (*ExtendedRemoteState
 }
 
 // Validate all the parameters of the given GCS remote state configuration
-func validateGCSConfig(extendedConfig *ExtendedRemoteStateConfigGCS, terragruntOptions *options.TerragruntOptions) error {
+func validateGCSConfig(extendedConfig *ExtendedRemoteStateConfigGCS) error {
 	var config = extendedConfig.remoteStateConfigGCS
 
-	if config.Prefix == "" {
-		return errors.WithStackTrace(MissingRequiredGCSRemoteStateConfig("prefix"))
+	if config.Bucket == "" {
+		return errors.WithStackTrace(MissingRequiredGCSRemoteStateConfig("bucket"))
 	}
 
 	return nil
@@ -271,10 +278,8 @@ func createGCSBucketIfNecessary(gcsClient *storage.Client, config *ExtendedRemot
 		if shouldCreateBucket {
 			// To avoid any eventual consistency issues with creating a GCS bucket we use a retry loop.
 			description := fmt.Sprintf("Create GCS bucket %s", config.remoteStateConfigGCS.Bucket)
-			maxRetries := 3
-			sleepBetweenRetries := 10 * time.Second
 
-			return util.DoWithRetry(description, maxRetries, sleepBetweenRetries, terragruntOptions.Logger, logrus.DebugLevel, func() error {
+			return util.DoWithRetry(description, gcpMaxRetries, gcpSleepBetweenRetries, terragruntOptions.Logger, logrus.DebugLevel, func() error {
 				return CreateGCSBucketWithVersioning(gcsClient, config, terragruntOptions)
 			})
 		}
@@ -470,9 +475,15 @@ func CreateGCSClient(gcsConfigRemote RemoteStateConfigGCS) (*storage.Client, err
 	}
 
 	if gcsConfigRemote.ImpersonateServiceAccount != "" {
-		opts = append(opts, option.ImpersonateCredentials(
-			gcsConfigRemote.ImpersonateServiceAccount,
-			gcsConfigRemote.ImpersonateServiceAccountDelegates...))
+		ts, err := impersonate.CredentialsTokenSource(ctx, impersonate.CredentialsConfig{
+			TargetPrincipal: gcsConfigRemote.ImpersonateServiceAccount,
+			Scopes:          []string{storage.ScopeFullControl},
+			Delegates:       gcsConfigRemote.ImpersonateServiceAccountDelegates,
+		})
+		if err != nil {
+			return nil, err
+		}
+		opts = append(opts, option.WithTokenSource(ts))
 	}
 
 	client, err := storage.NewClient(ctx, opts...)
