@@ -9,9 +9,9 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/aws_helper"
 	"github.com/gruntwork-io/terragrunt/dynamodb"
-	"github.com/gruntwork-io/terragrunt/errors"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/shell"
 	"github.com/gruntwork-io/terragrunt/util"
@@ -24,6 +24,10 @@ const (
 	DefaultS3BucketAccessLoggingTargetPrefix = "TFStateLogs/"
 	SidRootPolicy                            = "RootAccess"
 	SidEnforcedTLSPolicy                     = "EnforcedTLS"
+
+	s3TimeBetweenRetries  = 5 * time.Second
+	s3MaxRetries          = 3
+	s3SleepBetweenRetries = 10 * time.Second
 )
 
 /*
@@ -38,6 +42,7 @@ type ExtendedRemoteStateConfigS3 struct {
 	S3BucketTags                   map[string]string `mapstructure:"s3_bucket_tags"`
 	DynamotableTags                map[string]string `mapstructure:"dynamodb_table_tags"`
 	AccessLoggingBucketTags        map[string]string `mapstructure:"accesslogging_bucket_tags"`
+	SkipCredentialsValidation      bool              `mapstructure:"skip_credentials_validation"`
 	SkipBucketVersioning           bool              `mapstructure:"skip_bucket_versioning"`
 	SkipBucketSSEncryption         bool              `mapstructure:"skip_bucket_ssencryption"`
 	SkipBucketAccessLogging        bool              `mapstructure:"skip_bucket_accesslogging"`
@@ -159,6 +164,13 @@ func (s3Initializer S3Initializer) NeedsInitialization(remoteState *RemoteState,
 	s3Config := s3ConfigExtended.remoteStateConfigS3
 
 	sessionConfig := s3ConfigExtended.GetAwsSessionConfig()
+
+	// Validate current AWS session before checking S3
+	if !s3ConfigExtended.SkipCredentialsValidation {
+		if err = aws_helper.ValidateAwsSession(sessionConfig, terragruntOptions); err != nil {
+			return false, err
+		}
+	}
 
 	s3Client, err := CreateS3Client(sessionConfig, terragruntOptions)
 	if err != nil {
@@ -372,32 +384,39 @@ func validateS3Config(extendedConfig *ExtendedRemoteStateConfigS3, terragruntOpt
 // If the bucket specified in the given config doesn't already exist, prompt the user to create it, and if the user
 // confirms, create the bucket and enable versioning for it.
 func createS3BucketIfNecessary(s3Client *s3.S3, config *ExtendedRemoteStateConfigS3, terragruntOptions *options.TerragruntOptions) error {
-	if !DoesS3BucketExist(s3Client, &config.remoteStateConfigS3.Bucket) {
-		if terragruntOptions.FailIfBucketCreationRequired {
-			return BucketCreationNotAllowed(config.remoteStateConfigS3.Bucket)
-		}
+	if DoesS3BucketExist(s3Client, &config.remoteStateConfigS3.Bucket) {
+		return nil
+	}
+	if terragruntOptions.FailIfBucketCreationRequired {
+		return BucketCreationNotAllowed(config.remoteStateConfigS3.Bucket)
+	}
 
-		prompt := fmt.Sprintf("Remote state S3 bucket %s does not exist or you don't have permissions to access it. Would you like Terragrunt to create it?", config.remoteStateConfigS3.Bucket)
-		shouldCreateBucket, err := shell.PromptUserForYesNo(prompt, terragruntOptions)
-		if err != nil {
-			return err
-		}
+	prompt := fmt.Sprintf("Remote state S3 bucket %s does not exist or you don't have permissions to access it. Would you like Terragrunt to create it?", config.remoteStateConfigS3.Bucket)
+	shouldCreateBucket, err := shell.PromptUserForYesNo(prompt, terragruntOptions)
+	if err != nil {
+		return err
+	}
 
-		if shouldCreateBucket {
-			// Creating the S3 bucket occasionally fails with eventual consistency errors: e.g., the S3 HeadBucket
-			// operation says the bucket exists, but a subsequent call to enable versioning on that bucket fails with
-			// the error "NoSuchBucket: The specified bucket does not exist." Therefore, when creating and configuring
-			// the S3 bucket, we do so in a retry loop with a sleep between retries that will hopefully work around the
-			// eventual consistency issues. Each S3 operation should be idempotent, so redoing steps that have already
-			// been performed should be a no-op.
-			description := fmt.Sprintf("Create S3 bucket with retry %s", config.remoteStateConfigS3.Bucket)
-			maxRetries := 3
-			sleepBetweenRetries := 10 * time.Second
+	if shouldCreateBucket {
+		// Creating the S3 bucket occasionally fails with eventual consistency errors: e.g., the S3 HeadBucket
+		// operation says the bucket exists, but a subsequent call to enable versioning on that bucket fails with
+		// the error "NoSuchBucket: The specified bucket does not exist." Therefore, when creating and configuring
+		// the S3 bucket, we do so in a retry loop with a sleep between retries that will hopefully work around the
+		// eventual consistency issues. Each S3 operation should be idempotent, so redoing steps that have already
+		// been performed should be a no-op.
+		description := fmt.Sprintf("Create S3 bucket with retry %s", config.remoteStateConfigS3.Bucket)
 
-			return util.DoWithRetry(description, maxRetries, sleepBetweenRetries, terragruntOptions.Logger, logrus.DebugLevel, func() error {
-				return CreateS3BucketWithVersioningSSEncryptionAndAccessLogging(s3Client, config, terragruntOptions)
-			})
-		}
+		return util.DoWithRetry(description, s3MaxRetries, s3SleepBetweenRetries, terragruntOptions.Logger, logrus.DebugLevel, func() error {
+			err := CreateS3BucketWithVersioningSSEncryptionAndAccessLogging(s3Client, config, terragruntOptions)
+			if err != nil {
+				if isBucketCreationErrorRetriable(err) {
+					return err
+				}
+				// return FatalError so that retry loop will not continue
+				return util.FatalError{Underlying: err}
+			}
+			return nil
+		})
 	}
 
 	return nil
@@ -468,31 +487,9 @@ func updateS3BucketIfNecessary(s3Client *s3.S3, config *ExtendedRemoteStateConfi
 			terragruntOptions.Logger.Debugf("Access logging is disabled for the remote state AWS S3 bucket %s using 'skip_bucket_access_logging' config.", config.remoteStateConfigS3.Bucket)
 		} else {
 			if config.AccessLoggingBucketName != "" {
-				terragruntOptions.Logger.Debugf("Enabling bucket-wide Access Logging on AWS S3 bucket %s - using as TargetBucket %s", config.remoteStateConfigS3.Bucket, config.AccessLoggingBucketName)
-
-				if err := CreateLogsS3BucketIfNecessary(s3Client, aws.String(config.AccessLoggingBucketName), terragruntOptions); err != nil {
-					terragruntOptions.Logger.Errorf("Could not create logs bucket %s for AWS S3 bucket %s", config.AccessLoggingBucketName, config.remoteStateConfigS3.Bucket)
-					return err
+				if err := configureAccessLogBucket(terragruntOptions, s3Client, config); err != nil {
+					return nil
 				}
-
-				if err := EnablePublicAccessBlockingForS3Bucket(s3Client, config.AccessLoggingBucketName, terragruntOptions); err != nil {
-					return err
-				}
-
-				if err := EnableAccessLoggingForS3BucketWide(s3Client, &config.remoteStateConfigS3, terragruntOptions, config.AccessLoggingBucketName, config.AccessLoggingTargetPrefix); err != nil {
-					return err
-				}
-
-				if !config.SkipBucketSSEncryption {
-					if err := EnableSSEForS3BucketWide(s3Client, config.AccessLoggingBucketName, s3.ServerSideEncryptionAes256, config, terragruntOptions); err != nil {
-						return err
-					}
-				}
-
-				if err := EnableEnforcedTLSAccesstoS3Bucket(s3Client, config.AccessLoggingBucketName, config, terragruntOptions); err != nil {
-					return err
-				}
-
 			} else {
 				terragruntOptions.Logger.Debugf("Access Logging is disabled for the remote state AWS S3 bucket %s", config.remoteStateConfigS3.Bucket)
 			}
@@ -507,6 +504,35 @@ func updateS3BucketIfNecessary(s3Client *s3.S3, config *ExtendedRemoteStateConfi
 		}
 	}
 
+	return nil
+}
+
+// configureAccessLogBucket - configure access log bucket.
+func configureAccessLogBucket(terragruntOptions *options.TerragruntOptions, s3Client *s3.S3, config *ExtendedRemoteStateConfigS3) error {
+	terragruntOptions.Logger.Debugf("Enabling bucket-wide Access Logging on AWS S3 bucket %s - using as TargetBucket %s", config.remoteStateConfigS3.Bucket, config.AccessLoggingBucketName)
+
+	if err := CreateLogsS3BucketIfNecessary(s3Client, aws.String(config.AccessLoggingBucketName), terragruntOptions); err != nil {
+		terragruntOptions.Logger.Errorf("Could not create logs bucket %s for AWS S3 bucket %s\n%s", config.AccessLoggingBucketName, config.remoteStateConfigS3.Bucket, err.Error())
+		return errors.WithStackTrace(err)
+	}
+
+	if err := EnablePublicAccessBlockingForS3Bucket(s3Client, config.AccessLoggingBucketName, terragruntOptions); err != nil {
+		return errors.WithStackTrace(err)
+	}
+
+	if err := EnableAccessLoggingForS3BucketWide(s3Client, &config.remoteStateConfigS3, terragruntOptions, config.AccessLoggingBucketName, config.AccessLoggingTargetPrefix); err != nil {
+		return errors.WithStackTrace(err)
+	}
+
+	if !config.SkipBucketSSEncryption {
+		if err := EnableSSEForS3BucketWide(s3Client, config.AccessLoggingBucketName, s3.ServerSideEncryptionAes256, config, terragruntOptions); err != nil {
+			return errors.WithStackTrace(err)
+		}
+	}
+
+	if err := EnableEnforcedTLSAccesstoS3Bucket(s3Client, config.AccessLoggingBucketName, config, terragruntOptions); err != nil {
+		return errors.WithStackTrace(err)
+	}
 	return nil
 }
 
@@ -609,6 +635,8 @@ func checkIfS3BucketNeedsUpdate(s3Client *s3.S3, config *ExtendedRemoteStateConf
 
 // Check if versioning is enabled for the S3 bucket specified in the given config and warn the user if it is not
 func checkIfVersioningEnabled(s3Client *s3.S3, config *RemoteStateConfigS3, terragruntOptions *options.TerragruntOptions) (bool, error) {
+	terragruntOptions.Logger.Debugf("Verifying AWS S3 Bucket Versioning %s", config.Bucket)
+
 	out, err := s3Client.GetBucketVersioning(&s3.GetBucketVersioningInput{Bucket: aws.String(config.Bucket)})
 	if err != nil {
 		return false, errors.WithStackTrace(err)
@@ -626,9 +654,15 @@ func checkIfVersioningEnabled(s3Client *s3.S3, config *RemoteStateConfigS3, terr
 
 // Create the given S3 bucket and enable versioning for it
 func CreateS3BucketWithVersioningSSEncryptionAndAccessLogging(s3Client *s3.S3, config *ExtendedRemoteStateConfigS3, terragruntOptions *options.TerragruntOptions) error {
+	terragruntOptions.Logger.Debugf("Create S3 bucket %s with versioning, SSE encryption, and access logging.", config.remoteStateConfigS3.Bucket)
+
 	err := CreateS3Bucket(s3Client, aws.String(config.remoteStateConfigS3.Bucket), terragruntOptions)
 
 	if err != nil {
+		if accessError := checkBucketAccess(s3Client, aws.String(config.remoteStateConfigS3.Bucket), aws.String(config.remoteStateConfigS3.Key)); accessError != nil {
+			return accessError
+		}
+
 		if isBucketAlreadyOwnedByYouError(err) {
 			terragruntOptions.Logger.Debugf("Looks like you're already creating bucket %s at the same time. Will not attempt to create it again.", config.remoteStateConfigS3.Bucket)
 			return WaitUntilS3BucketExists(s3Client, &config.remoteStateConfigS3, terragruntOptions)
@@ -680,29 +714,8 @@ func CreateS3BucketWithVersioningSSEncryptionAndAccessLogging(s3Client *s3.S3, c
 	}
 
 	if config.AccessLoggingBucketName != "" {
-		terragruntOptions.Logger.Debugf("Enabling bucket-wide Access Logging on AWS S3 bucket %s - using as TargetBucket %s", config.remoteStateConfigS3.Bucket, config.AccessLoggingBucketName)
-
-		if err := CreateLogsS3BucketIfNecessary(s3Client, aws.String(config.AccessLoggingBucketName), terragruntOptions); err != nil {
-			terragruntOptions.Logger.Errorf("Could not create logs bucket %s for AWS S3 bucket %s", config.AccessLoggingBucketName, config.remoteStateConfigS3.Bucket)
-			return err
-		}
-
-		if err := EnablePublicAccessBlockingForS3Bucket(s3Client, config.AccessLoggingBucketName, terragruntOptions); err != nil {
-			return err
-		}
-
-		if err := EnableAccessLoggingForS3BucketWide(s3Client, &config.remoteStateConfigS3, terragruntOptions, config.AccessLoggingBucketName, config.AccessLoggingTargetPrefix); err != nil {
-			return err
-		}
-
-		if !config.SkipBucketSSEncryption {
-			if err := EnableSSEForS3BucketWide(s3Client, config.AccessLoggingBucketName, s3.ServerSideEncryptionAes256, config, terragruntOptions); err != nil {
-				return err
-			}
-		}
-
-		if err := EnableEnforcedTLSAccesstoS3Bucket(s3Client, config.AccessLoggingBucketName, config, terragruntOptions); err != nil {
-			return err
+		if err := configureAccessLogBucket(terragruntOptions, s3Client, config); err != nil {
+			return nil
 		}
 	} else {
 		terragruntOptions.Logger.Debugf("Access Logging is disabled for the remote state AWS S3 bucket %s", config.remoteStateConfigS3.Bucket)
@@ -840,6 +853,15 @@ func isBucketAlreadyOwnedByYouError(err error) bool {
 	return isAwsErr && (awsErr.Code() == "BucketAlreadyOwnedByYou" || awsErr.Code() == "OperationAborted")
 }
 
+// isBucketCreationErrorRetriable returns true if the error is temporary and bucket creation can be retried.
+func isBucketCreationErrorRetriable(err error) bool {
+	awsErr, isAwsErr := errors.Unwrap(err).(awserr.Error)
+	if !isAwsErr {
+		return true
+	}
+	return awsErr.Code() == "InternalError" || awsErr.Code() == "OperationAborted" || awsErr.Code() == "InvalidParameter"
+}
+
 // Add a policy to allow root access to the bucket
 func EnableRootAccesstoS3Bucket(s3Client *s3.S3, config *ExtendedRemoteStateConfigS3, terragruntOptions *options.TerragruntOptions) error {
 	bucket := config.remoteStateConfigS3.Bucket
@@ -927,7 +949,7 @@ func checkIfBucketRootAccess(s3Client *s3.S3, config *RemoteStateConfigS3, terra
 	})
 	if err != nil {
 		terragruntOptions.Logger.Debugf("Could not get policy for bucket %s", config.Bucket)
-		return false, nil
+		return false, err
 	}
 
 	// If the bucket has no policy, it is not enforced
@@ -1136,7 +1158,7 @@ func checkIfSSEForS3Enabled(s3Client *s3.S3, config *ExtendedRemoteStateConfigS3
 	output, err := s3Client.GetBucketEncryption(input)
 	if err != nil {
 		terragruntOptions.Logger.Debugf("Error checking if SSE is enabled for AWS S3 bucket %s: %s", config.remoteStateConfigS3.Bucket, err.Error())
-		return false, nil
+		return false, err
 	}
 
 	if output.ServerSideEncryptionConfiguration == nil {
@@ -1191,7 +1213,7 @@ func checkIfAccessLoggingForS3Enabled(s3Client *s3.S3, config *RemoteStateConfig
 	output, err := s3Client.GetBucketLogging(input)
 	if err != nil {
 		terragruntOptions.Logger.Debugf("Error checking if Access Logging is enabled for AWS S3 bucket %s: %s", config.Bucket, err.Error())
-		return false, nil
+		return false, err
 	}
 
 	if output.LoggingEnabled == nil {
@@ -1228,6 +1250,7 @@ func EnablePublicAccessBlockingForS3Bucket(s3Client *s3.S3, bucketName string, t
 
 func checkIfS3PublicAccessBlockingEnabled(s3Client *s3.S3, config *RemoteStateConfigS3, terragruntOptions *options.TerragruntOptions) (bool, error) {
 	terragruntOptions.Logger.Debugf("Checking if S3 bucket %s is configured to block public access", config.Bucket)
+
 	output, err := s3Client.GetPublicAccessBlock(&s3.GetPublicAccessBlockInput{
 		Bucket: aws.String(config.Bucket),
 	})
@@ -1291,7 +1314,6 @@ func waitUntilBucketHasAccessLoggingAcl(s3Client *s3.S3, bucket *string, terragr
 	terragruntOptions.Logger.Debugf("Waiting for ACL bucket %s to have the updated ACL for access logging.", aws.StringValue(bucket))
 
 	maxRetries := 10
-	timeBetweenRetries := 5 * time.Second
 
 	for i := 0; i < maxRetries; i++ {
 		out, err := s3Client.GetBucketAcl(&s3.GetBucketAclInput{Bucket: bucket})
@@ -1318,8 +1340,8 @@ func waitUntilBucketHasAccessLoggingAcl(s3Client *s3.S3, bucket *string, terragr
 			return nil
 		}
 
-		terragruntOptions.Logger.Debugf("Bucket %s still does not have the ACL permissions for access logging. Will sleep for %v and check again.", aws.StringValue(bucket), timeBetweenRetries)
-		time.Sleep(timeBetweenRetries)
+		terragruntOptions.Logger.Debugf("Bucket %s still does not have the ACL permissions for access logging. Will sleep for %v and check again.", aws.StringValue(bucket), s3TimeBetweenRetries)
+		time.Sleep(s3TimeBetweenRetries)
 	}
 
 	return errors.WithStackTrace(MaxRetriesWaitingForS3ACLExceeded(aws.StringValue(bucket)))
@@ -1330,6 +1352,23 @@ func waitUntilBucketHasAccessLoggingAcl(s3Client *s3.S3, bucket *string, terragr
 func DoesS3BucketExist(s3Client *s3.S3, bucket *string) bool {
 	_, err := s3Client.HeadBucket(&s3.HeadBucketInput{Bucket: bucket})
 	return err == nil
+}
+
+// checkBucketAccess checks if the current user has the ability to access the S3 bucket keys.
+func checkBucketAccess(s3Client *s3.S3, bucket *string, key *string) error {
+	_, err := s3Client.GetObject(&s3.GetObjectInput{Key: key, Bucket: bucket})
+	if err == nil {
+		return nil
+	}
+	awsErr, isAwsErr := errors.Unwrap(err).(awserr.Error)
+	if !isAwsErr {
+		return err
+	}
+	// filter permissions errors
+	if awsErr.Code() == "NoSuchBucket" || awsErr.Code() == "NoSuchKey" {
+		return nil
+	}
+	return err
 }
 
 // Create a table for locks in DynamoDB if the user has configured a lock table and the table doesn't already exist
