@@ -261,17 +261,19 @@ func TestGetTerraformInitArgs(t *testing.T) {
 		{
 			"terragrunt-keys-filtered",
 			map[string]interface{}{
-				"bucket":         "foo",
-				"encrypt":        "bar",
-				"key":            "baz",
-				"region":         "quux",
-				"s3_bucket_tags": map[string]string{},
+				"bucket":                      "foo",
+				"encrypt":                     "bar",
+				"key":                         "baz",
+				"region":                      "quux",
+				"skip_credentials_validation": true,
+				"s3_bucket_tags":              map[string]string{},
 			},
 			map[string]interface{}{
-				"bucket":  "foo",
-				"encrypt": "bar",
-				"key":     "baz",
-				"region":  "quux",
+				"bucket":                      "foo",
+				"encrypt":                     "bar",
+				"key":                         "baz",
+				"region":                      "quux",
+				"skip_credentials_validation": true,
 			},
 			true,
 		},
