@@ -8,11 +8,17 @@ import (
 	"os/exec"
 	"os/signal"
 	"strings"
+	"sync"
 	"syscall"
 	"time"
 
-	"github.com/gruntwork-io/terragrunt/errors"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
+
+	"github.com/gruntwork-io/go-commons/collections"
+	"github.com/gruntwork-io/go-commons/errors"
 	"github.com/gruntwork-io/terragrunt/options"
+	"github.com/gruntwork-io/terragrunt/terraform"
 	"github.com/gruntwork-io/terragrunt/util"
 	"github.com/hashicorp/go-multierror"
 	"github.com/sirupsen/logrus"
@@ -31,9 +37,16 @@ var terraformCommandsThatNeedPty = []string{
 	"console",
 }
 
+var terraformInitMutex sync.Mutex
+
 // Run the given Terraform command
 func RunTerraformCommand(terragruntOptions *options.TerragruntOptions, args ...string) error {
-	_, err := RunShellCommandWithOutput(terragruntOptions, "", false, isTerraformCommandThatNeedsPty(args[0]), terragruntOptions.TerraformPath, args...)
+	needPTY, err := isTerraformCommandThatNeedsPty(args)
+	if err != nil {
+		return err
+	}
+
+	_, err = RunShellCommandWithOutput(terragruntOptions, "", false, needPTY, terragruntOptions.TerraformPath, args...)
 	return err
 }
 
@@ -46,11 +59,12 @@ func RunShellCommand(terragruntOptions *options.TerragruntOptions, command strin
 // Run the given Terraform command, writing its stdout/stderr to the terminal AND returning stdout/stderr to this
 // method's caller
 func RunTerraformCommandWithOutput(terragruntOptions *options.TerragruntOptions, args ...string) (*CmdOutput, error) {
-	needPty := false
-	if len(args) > 0 {
-		needPty = isTerraformCommandThatNeedsPty(args[0])
+	needPTY, err := isTerraformCommandThatNeedsPty(args)
+	if err != nil {
+		return nil, err
 	}
-	return RunShellCommandWithOutput(terragruntOptions, "", false, needPty, terragruntOptions.TerraformPath, args...)
+
+	return RunShellCommandWithOutput(terragruntOptions, "", false, needPTY, terragruntOptions.TerraformPath, args...)
 }
 
 // Run the specified shell command with the specified arguments. Connect the command's stdin, stdout, and stderr to
@@ -64,6 +78,14 @@ func RunShellCommandWithOutput(
 	command string,
 	args ...string,
 ) (*CmdOutput, error) {
+	// Terrafrom `init` command with the plugin cache directory is not guaranteed to be concurrency safe.
+	// The provider installer's behavior in environments with multiple terraform init calls is undefined.
+	// Thus, terraform `init` commands must be executed sequentially, even if `--terragrunt-parallelism` is greater than 1.
+	if command == "terraform" && collections.ListContainsElement(args, "init") && terraform.IsPluginCacheUsed() {
+		defer terraformInitMutex.Unlock()
+		terraformInitMutex.Lock()
+	}
+
 	terragruntOptions.Logger.Debugf("Running command: %s %s", command, strings.Join(args, " "))
 	if suppressStdout {
 		terragruntOptions.Logger.Debugf("Command output will be suppressed.")
@@ -118,7 +140,12 @@ func RunShellCommandWithOutput(
 	// Make sure to forward signals to the subcommand.
 	cmdChannel := make(chan error) // used for closing the signals forwarder goroutine
 	signalChannel := NewSignalsForwarder(forwardSignals, cmd, terragruntOptions.Logger, cmdChannel)
-	defer signalChannel.Close()
+	defer func(signalChannel *SignalsForwarder) {
+		err := signalChannel.Close()
+		if err != nil {
+			terragruntOptions.Logger.Warnf("Error closing signal channel: %v", err)
+		}
+	}(&signalChannel)
 
 	err := cmd.Wait()
 	cmdChannel <- err
@@ -149,14 +176,33 @@ func toEnvVarsList(envVarsAsMap map[string]string) []string {
 }
 
 // isTerraformCommandThatNeedsPty returns true if the sub command of terraform we are running requires a pty.
-func isTerraformCommandThatNeedsPty(command string) bool {
-	return util.ListContainsElement(terraformCommandsThatNeedPty, command)
+func isTerraformCommandThatNeedsPty(args []string) (bool, error) {
+	if len(args) == 0 || !util.ListContainsElement(terraformCommandsThatNeedPty, args[0]) {
+		return false, nil
+	}
+
+	fi, err := os.Stdin.Stat()
+	if err != nil {
+		return false, errors.WithStackTrace(err)
+	}
+
+	// if there is data in the stdin, then the terraform console is used in non-interactive mode, for example `echo "1 + 5" | terragrunt console`.
+	if fi.Size() > 0 {
+		return false, nil
+	}
+
+	return true, nil
 }
 
-// Return the exit code of a command. If the error does not implement errors.IErrorCode or is not an exec.ExitError
+// Return the exit code of a command. If the error does not implement iErrorCode or is not an exec.ExitError
 // or *multierror.Error type, the error is returned.
 func GetExitCode(err error) (int, error) {
-	if exiterr, ok := errors.Unwrap(err).(errors.IErrorCode); ok {
+	// Interface to determine if we can retrieve an exit status from an error
+	type iErrorCode interface {
+		ExitStatus() (int, error)
+	}
+
+	if exiterr, ok := errors.Unwrap(err).(iErrorCode); ok {
 		return exiterr.ExitStatus()
 	}
 
@@ -196,7 +242,7 @@ func NewSignalsForwarder(signals []os.Signal, c *exec.Cmd, logger *logrus.Entry,
 		for {
 			select {
 			case s := <-signalChannel:
-				logger.Debugf("%s signal received. Gracefully shutting down... (it can take up to %v)", strings.Title(s.String()), signalForwardingDelay)
+				logger.Debugf("%s signal received. Gracefully shutting down... (it can take up to %v)", cases.Title(language.English).String(s.String()), signalForwardingDelay)
 
 				select {
 				case <-time.After(signalForwardingDelay):
@@ -234,7 +280,7 @@ type CmdOutput struct {
 func GitTopLevelDir(terragruntOptions *options.TerragruntOptions, path string) (string, error) {
 	stdout := bytes.Buffer{}
 	stderr := bytes.Buffer{}
-	opts, err := options.NewTerragruntOptions(path)
+	opts, err := options.NewTerragruntOptionsWithConfigPath(path)
 	if err != nil {
 		return "", err
 	}
@@ -242,7 +288,7 @@ func GitTopLevelDir(terragruntOptions *options.TerragruntOptions, path string) (
 	opts.Writer = &stdout
 	opts.ErrWriter = &stderr
 	cmd, err := RunShellCommandWithOutput(opts, path, true, false, "git", "rev-parse", "--show-toplevel")
-	terragruntOptions.Logger.Debugf("git show-toplevel result: \n%v\n%v\n", (string)(stdout.Bytes()), (string)(stderr.Bytes()))
+	terragruntOptions.Logger.Debugf("git show-toplevel result: \n%v\n%v\n", stdout.String(), stderr.String())
 	if err != nil {
 		return "", err
 	}
