@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"regexp"
 	"strings"
-
-	"github.com/gruntwork-io/terragrunt/errors"
 )
 
 func MatchesAny(regExps []string, s string) bool {
@@ -119,10 +117,8 @@ func CommaSeparatedStrings(list []string) string {
 
 // Make a copy of the given list of strings
 func CloneStringList(listToClone []string) []string {
-	out := []string{}
-	for _, item := range listToClone {
-		out = append(out, item)
-	}
+	var out []string
+	out = append(out, listToClone...)
 	return out
 }
 
@@ -169,27 +165,6 @@ func StringListInsert(list []string, element string, index int) []string {
 	return append(list[:index], tail...)
 }
 
-// KeyValuePairListToMap converts a list of key value pair encoded as `key=value` strings into a map
-// using the given `splitter` callback func, which can be the `strings.Split` function.
-func KeyValuePairStringListToMap(asList []string, splitter func(s, sep string) []string) (map[string]string, error) {
-	asMap := map[string]string{}
-
-	for _, arg := range asList {
-		parts := splitter(arg, "=")
-
-		if len(parts) != 2 {
-			return nil, errors.WithStackTrace(InvalidKeyValue(arg))
-		}
-
-		key := parts[0]
-		value := parts[1]
-
-		asMap[key] = value
-	}
-
-	return asMap, nil
-}
-
 // SplitUrls slices s into all substrings separated by sep and returns a slice of
 // the substrings between those separators.
 // Taking into account that the `=` sign can also be used as a git tag, e.g. `git@github.com/test.git?ref=feature`
@@ -200,7 +175,7 @@ func SplitUrls(s, sep string) []string {
 
 	// mask
 	for src, mask := range masks {
-		s = strings.Replace(s, src, mask, -1)
+		s = strings.ReplaceAll(s, src, mask)
 	}
 
 	urls := strings.Split(s, sep)
@@ -208,17 +183,9 @@ func SplitUrls(s, sep string) []string {
 	// unmask
 	for i := range urls {
 		for src, mask := range masks {
-			urls[i] = strings.Replace(urls[i], mask, src, -1)
+			urls[i] = strings.ReplaceAll(urls[i], mask, src)
 		}
 	}
 
 	return urls
-}
-
-// custom error types
-
-type InvalidKeyValue string
-
-func (err InvalidKeyValue) Error() string {
-	return fmt.Sprintf("Invalid key-value pair. Expected format KEY=VALUE, got %s.", string(err))
 }
